Update words.php
I have translated all the words in this file, there are definitely errors, but I do not know how it looks in the game, how could I see them? I'm playing very good, I've been playing for a few days and it really is great. I will try to translate on a regular basis.
Regards Philip